+++ src/combinator/ops/mul.rs
@@ -64,7 +64,7 @@
 //!
 //! // parse "123" to 123
 //! assert_eq!(
-//!   combinator.exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()).unwrap().value,
+//!   combinator.exec(Input::new("123", 0, &mut (), &mut ()).unwrap()).unwrap().value,
 //!   123
 //! )
 //! ```
@@ -95,7 +95,7 @@
 //!
 //! // parse "123" to 123
 //! assert_eq!(
-//!   combinator.exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()).unwrap().value,
+//!   combinator.exec(Input::new("123", 0, &mut (), &mut ()).unwrap()).unwrap().value,
 //!   123
 //! )
 //! ```
@@ -151,13 +151,13 @@ mod tests {
 
     // repeat a rejecter will reject
     assert!((rejecter() * 3)
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
 
     // repeat rejecter 0 times will accept
     let n = 0;
     assert_eq!(
-      (rejecter() * n).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * n).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -167,7 +167,7 @@ mod tests {
     // repeat an accepter 0 times will accept
     let n = 0;
     assert_eq!(
-      (accepter() * n).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * n).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 0,
         digested: 0,
@@ -176,7 +176,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * 3).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * 3).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 3,
         digested: 3
@@ -185,7 +185,7 @@ mod tests {
 
     // overflow, reject
     assert!((accepter() * 4)
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
   }
 
@@ -202,12 +202,12 @@ mod tests {
 
     // repeat a rejecter will reject
     assert!((rejecter() * (1..2))
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
 
     // repeat rejecter 0 times will accept
     assert_eq!(
-      (rejecter() * (0..2)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * (0..2)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -216,7 +216,7 @@ mod tests {
 
     // repeat an accepter 0 times will accept
     assert_eq!(
-      (accepter() * (0..1)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (0..1)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 0,
         digested: 0,
@@ -225,7 +225,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * (0..3)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (0..3)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 1,
         digested: 2
@@ -234,7 +234,7 @@ mod tests {
 
     // too few, reject
     assert!((accepter() * (4..6))
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
   }
 
@@ -251,12 +251,12 @@ mod tests {
 
     // repeat a rejecter will reject
     assert!((rejecter() * (1..))
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
 
     // repeat rejecter 0 times will accept
     assert_eq!(
-      (rejecter() * (0..)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * (0..)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -265,7 +265,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * (0..)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (0..)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 3,
         digested: 3
@@ -274,7 +274,7 @@ mod tests {
 
     // too few, reject
     assert!((accepter() * (4..))
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
   }
 
@@ -291,7 +291,7 @@ mod tests {
 
     // repeat rejecter 0 times will accept
     assert_eq!(
-      (rejecter() * (..)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * (..)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -300,7 +300,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * (..)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (..)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 3,
         digested: 3
@@ -321,12 +321,12 @@ mod tests {
 
     // repeat a rejecter will reject
     assert!((rejecter() * (1..=3))
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
 
     // repeat rejecter 0 times will accept
     assert_eq!(
-      (rejecter() * (0..=2)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * (0..=2)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -335,7 +335,7 @@ mod tests {
 
     // repeat an accepter 0 times will accept
     assert_eq!(
-      (accepter() * (0..=0)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (0..=0)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 0,
         digested: 0,
@@ -344,7 +344,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * (0..=3)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (0..=3)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 3,
         digested: 3
@@ -353,7 +353,7 @@ mod tests {
 
     // too few, reject
     assert!((accepter() * (4..=6))
-      .exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap())
+      .exec(Input::new("123", 0, &mut (), &mut ()).unwrap())
       .is_none());
   }
 
@@ -370,7 +370,7 @@ mod tests {
 
     // repeat rejecter 0 times will accept
     assert_eq!(
-      (rejecter() * (..2)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * (..2)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -379,7 +379,7 @@ mod tests {
 
     // repeat an accepter 0 times will accept
     assert_eq!(
-      (accepter() * (..1)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (..1)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 0,
         digested: 0,
@@ -388,7 +388,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * (..3)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (..3)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 1,
         digested: 2
@@ -409,7 +409,7 @@ mod tests {
 
     // repeat rejecter 0 times will accept
     assert_eq!(
-      (rejecter() * (..=2)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (rejecter() * (..=2)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: (),
         digested: 0,
@@ -418,7 +418,7 @@ mod tests {
 
     // repeat an accepter 0 times will accept
     assert_eq!(
-      (accepter() * (..=0)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (..=0)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 0,
         digested: 0,
@@ -427,7 +427,7 @@ mod tests {
 
     // normal, apply the folded value and sum the digested
     assert_eq!(
-      (accepter() * (..=3)).exec(&mut Input::new("123", 0, &mut (), &mut ()).unwrap()),
+      (accepter() * (..=3)).exec(Input::new("123", 0, &mut (), &mut ()).unwrap()),
       Some(Output {
         value: 3,
         digested: 3